+++ CHANGELOG.md
@@ -6,5 +6,5 @@
 
 ### 7.0.0
 
-- 【版本】升级 7.0.0，适用于 angular 7, ng-zorro-antd 7 版本
+- 【版本】升级 7.0.0，适用于 angular 7, ng-zorro-antd 7.5.0 版本
 - 【hn-preview】修复 hn-preview 组件左右箭头不显示的 bug